Refactor: Translate scripts to English and update README
This commit refactors the shell scripts by translating them to English.
The scripts were renamed to include "_en" suffix.
Also, the README.md file was updated to reflect these changes.
This improves the accessibility and maintainability of the scripts for a wider audience.
The scripts 6_snap-apps.sh and 8_zerotier-client.sh were removed and replaced with their english versions.